Usage as a module
Having a class including `Kwork` transparently creates a
`Kwork::Transaction` instance under the hood. The class gets a
`#transaction` method which is delegated to the transaction instance.

Operations and adapter are specified when including the module through
the `.[]` method arguments.

The including class can automatically delegate to the executor through
the `MethodMissing` module.