Allow using including instance methods as operations
When the operations are resolved from the including class, they're
resolved to instance methods when they're given as symbols.

That forced us to remove the option to invisibly delegate to the runner.
As the operation would share the name with the instance method name,
Ruby would always call the instance method instead of the operation,
missing the unwrapping of the result performed by the runner.

The other option would be to change the context within the block to be
the runner instead of the instance (`instance_eval`), but that would
make other instance methods unavailable, unless we found a way to
delegate missing methods to the instance.

However, after consideration, we reasoned that the transparent
delegation was an anti-pattern, as one could want to call one of the
operations and not want it to be unwrapped by the runner (e.g. with
dry-monads, `perform_something(x).or(false)`).