Add Kwork::Result#bind
The difference between #bind and #map is that the former expects the
given block to return a Kwork::Result, while the latter performs the
wrapping automatically.

Given that Ruby is a dynamic language, it is not possible to enforce
the block's type signature. However, that's the best match we can get with
what, for instance, Haskell provides:

>>= :: m a -> (a -> m b) -> m b
fmap :: (a -> b) -> f a -> f b