Use `step operation` instead of `t.operation`
This makes the feel of the DSL more natural.

It's important to note that's changing how profiling works, as well as
the internal implementation of the `transaction` block.

When the Ruby interpreter sees:

```ruby
step operation
```

It will first call the `operation` method. As we need to decorate it to
throw `:halt` in case of failure (previously, that was done via the
magic method defined in the runner: `t.operation`), we changed the context
of the block to be the runner itself. If previously, only calls on `t.`
were profiled, now all calls to instance methods are profiled:

Before:

```ruby
operation # operation won't be profiled
t.operation # operation will be profiled
```

Now:

```ruby
operation # operation will be profiled
t.operation # operation will be profiled
```

That's not bad per se, as it gives more information, considering
everything within the transaction block as steps.

What's improved now is that we can make a `step` from an external call:

```ruby
step SomeClass.some_operation
```

Although the latter won't go through the profiler.